+++ apis/couple.ts
@@ -1,62 +1,4 @@
-import LocalStorage from 'utils/storage';
 import { AxiosInstance } from 'axios';
-import { SetterOrUpdater } from 'recoil';
-import { UserResponseType } from 'types/auth';
-
-interface IUploadUserProps {
-  instance: AxiosInstance;
-  accessToken: string;
-  setUser: SetterOrUpdater<UserResponseType | null>;
-}
-export async function updateUserInfoWhenCoupleLinked({
-  instance,
-  accessToken,
-  setUser,
-}: IUploadUserProps) {
-  const res = instance
-    .get('/members')
-    .then((response) => {
-      return response.data;
-    })
-    .catch((error) => {
-      console.error(error);
-      return {
-        data: null,
-      };
-    });
-
-  const { data } = await res;
-
-  if (data) {
-    LocalStorage.setItem('accessToken', accessToken);
-    setUser(data);
-  }
-}
-
-export function updateMemberInfo({
-  instance,
-  data,
-}: {
-  instance: AxiosInstance;
-  data: {
-    imageUrl: string;
-    nickName: string;
-  };
-}) {
-  const res = instance
-    .put('/members', data)
-    .then((response) => {
-      return response.data;
-    })
-    .catch((error) => {
-      console.error(error);
-      return {
-        data: null,
-      };
-    });
-
-  return res;
-}
 
 export function updateCoupleInfo({
   instance,
