+++ components/organisms/ProfileEditForm/ProfileEditForm.tsx
@@ -5,7 +5,8 @@ import { ProfileUpload, Button, SubmitButton } from 'components';
 import { useImage, useForm, useAxiosInstance } from 'hooks';
 import { IUserProps, IEditState, EditErrorTypes, IOnSubmit } from 'types';
 import { UserResponseType } from 'types/auth';
-import { updateMemberInfo, updateCoupleInfo } from 'apis/couple';
+import { updateCoupleInfo } from 'apis/couple';
+import { updateMemberInfo } from 'apis/members';
 import { profileEditValidation } from 'utils/formValidation';
 import { CoupleForm } from './CoupleForm';
 import { SoloForm } from './SoloForm';
